pyproject.toml
[project]
name = "omaha-oracle"
version = "0.1.0"
description = "Omaha Oracle — Autonomous value investing agent built on Graham-Dodd-Buffett principles"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.34",
    "requests>=2.31",
    "yfinance>=0.2.36",
    "anthropic>=0.40",
    "alpaca-py>=0.25",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.28",
    "pandas>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "moto>=5.0",
    "ruff>=0.4",
    "mypy>=1.9",
    "boto3-stubs[essential]>=1.34",
]
infra = [
    "aws-cdk-lib>=2.130",
    "constructs>=10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.mypy]
python_version = "3.12"
strict = true